middleware::log_request: Split request timing code into dedicated middleware
